Separate cluster.py in management and handler submodules
